+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [Next release]
 
 ### ✨ Added
+- `StratifiedClassicalMeanEstimator` for Stratified classical mean estimation
+- Export JSON step in quickstart
 - Stratified PPI tutorial notebook
 - Code of conduct
 - Tutorial for ASI